MDEV-16546 System versioning setting to allow history modification
Add server variable system_versioning_modify_history which will allow
to modify history by DML operations.

If secure_timestamp is YES or REPLICATION or execution is from
prepared statement system_versioning_modify_history does not have
effect. If secure_timestamp is SUPER, system_versioning_modify_history
requires special privilege (same as for setting current timestamp).

Setting system_versioning_modify_history to ON is disallowed if
secure_timestamp is ON or REPLICATION. If secure_timestamp is SUPER,
allow only for superuser. This is applicable to command-line as well.

UPDATE, DELETE of historical rows is allowed only for superuser and
only when row_start or row_end is specified explicitly in WHERE
condition.

Note that PS is disabled because permanent Item::transform() is not
supported which is the subject for future refactoring if it will be
needed.

Depends on:

MDEV-18727 improve DML operation of System Versioning

[Closes tempesta-tech/mariadb#592]