MDEV-21452 Default MUTEXTYPE=sys for improved performance
On POWER and ARMv8, the native mutex implementation of the standard
library on GNU/Linux seems to perform noticeably better than
the current default MUTEXTYPE=event or the MUTEXTYPE=futex.

On x86, the difference is smaller, and the MUTEXTYPE=event might be
slightly faster than the system mutex, but not by much.

Let us default to MUTEXTYPE=sys, that is, sys_mutex_t
(pthread_mutex_t or CRITICAL_SECTION) wrapped by
OSMutex and OSTrackMutex. Later, we might remove the
other implementations altogether.