MDEV-22441 implement a generic way to change a value of a variable in scope
Scope_value<T> class changes a variable value to some temporary value while
it exits and restores an old value in it's destructor.

make_scope_value() is a function targeted for C++11 which allows to omit
a template parameter which is required by Scope_value<T>

Sql_mode_save removed and the only usage is replaced with Scope_value<T>

Sql_mode_save removal should not be merged in newer branches!