MDEV-22583: Selectivity for BIT columns in filtered column for EXPLAIN is incorrect
For BIT columns when EITS is collected, we store the integral value in text representation
in the min and max fields of the statistical table (that is mysql.column_stats)
When this value is retrieved from the statistical table to original table field
then we try to store the text representation in the original field which is INCORRECT.
The value that is retrieved should be converted to integral type and that value should
be stored back in the original field. This would get us the correct estimate for
selectivity of the predicate.